feat :: navigate to sign in function
