+++ app/src/main/java/team/retum/jobisandroidv2/navigation/AuthNavigation.kt
@@ -1,17 +1,21 @@
 package team.retum.jobisandroidv2.navigation
 
+import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.navigation
 import team.retum.landing.NAVIGATION_LANDING
 import team.retum.landing.landing
+import team.retum.signin.navigateToSignIn
+import team.retum.signin.signIn
 
 const val NAVIGATION_AUTH = "auth"
 
-internal fun NavGraphBuilder.authNavigation() {
+internal fun NavGraphBuilder.authNavigation(navController: NavController) {
     navigation(
         route = NAVIGATION_AUTH,
         startDestination = NAVIGATION_LANDING,
     ) {
-        landing()
+        landing(onSignInClick = navController::navigateToSignIn)
+        signIn()
     }
 }
